refactor: 피드백 반영
- Entity에 Optional=false 추가
- converter 패키지 변경
